export request from index
