+++ src/utils/request/index.js
@@ -1,12 +1,2 @@
-import checkStatus from './checkStatus';
-import applyPlugins from '../applyPlugins';
-
-const request = (url, fetchOptions = {}, configuration = {}) => {
-  const { plugins: { response: responsePlugins = [] } = {} } = configuration;
-
-  return fetch(url, fetchOptions)
-    .then(checkStatus)
-    .then(data => applyPlugins(data, responsePlugins));
-};
-
+import request from './request';
 export default request;
